+++ cedar-policy/CHANGELOG.md
@@ -35,6 +35,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Improve parse error around invalid `is` expressions.
 - Improve parser error message when a policy includes an invalid template slot.
   The error now identifies that the policy used an invalid slot and suggests using
   one of the valid slots.
@@ -107,6 +108,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Improve schema parsing error messages when a cycle exists in the action
   hierarchy to includes an action which is part of the cycle (#436, resolving
   #416).
+- For the `partial-eval` experimental feature: make the return values of
+  `RequestBuilder`'s `principal`, `action`, `resource`, `context` and
+  `schema` functions `#[must_use]`.
+- Add PolicyId `AsRef<str>` implementation.
 
 ### Fixed